build: breakdown gz archive task
